remove some explicit type casts
implicit type casts should work here as expected

+++ w-image-viewer/user_interface.cpp
@@ -377,11 +377,11 @@ void User_interface::auto_window_size() const
 	//if the image resolution is larger than the screen resolution * 0.9, downsize the window to screen resolution * 0.9 with the aspect ratio of the image
 	if (cx_screen / cy_screen > image_width / image_height) {
 		rect.right = image_width > cx_screen * 0.9 ? std::lround(image_width * cy_screen / image_height * 0.9) : static_cast<LONG>(image_width);
-		rect.bottom = image_height > cy_screen * 0.9 ? std::lround(image_height * static_cast<double>(rect.right) / image_width) : static_cast<LONG>(image_height);
+		rect.bottom = image_height > cy_screen * 0.9 ? std::lround(image_height * rect.right / image_width) : static_cast<LONG>(image_height);
 	}
 	else {
 		rect.bottom = image_height > cy_screen * 0.9 ? std::lround(image_height * cx_screen / image_width * 0.9) : static_cast<LONG>(image_height);
-		rect.right = image_width > cx_screen * 0.9 ? std::lround(image_width * static_cast<double>(rect.bottom) / image_height) : static_cast<LONG>(image_width);
+		rect.right = image_width > cx_screen * 0.9 ? std::lround(image_width * rect.bottom / image_height) : static_cast<LONG>(image_width);
 	}
 
 	AdjustWindowRectEx(&rect, WIV_WINDOW_STYLE, FALSE, WIV_WINDOW_EX_STYLE);
@@ -391,7 +391,7 @@ void User_interface::auto_window_size() const
 	const auto cy{ rect.bottom - rect.top };
 
 	//center the window and apply new dimensions
-	SetWindowPos(hwnd, nullptr, static_cast<int>((cx_screen - static_cast<double>(cx)) / 2.0), static_cast<int>((cy_screen - static_cast<double>(cy)) / 2.0), cx, cy, SWP_NOZORDER);
+	SetWindowPos(hwnd, nullptr, static_cast<int>((cx_screen - cx) / 2.0), static_cast<int>((cy_screen - cy) / 2.0), cx, cy, SWP_NOZORDER);
 }
 
 //imgui dimming helper
